feat: 가계부 추가 모달에서 Category 리스트 가져오기
