+++ client/src/api/categoryAPI.ts
@@ -1,0 +1,6 @@
+import { Category } from "../interfaces/Category";
+import { getFetch, Result } from "./fetch";
+
+export const getCategoriesAsync = async (): Promise<Result<Category[]>> => {
+    return await getFetch("/category");
+};
