+++ client/src/interfaces/Category.ts
@@ -1,4 +1,5 @@
-export interface ICategoryItem {
+export interface Category {
   id: number;
   name: string;
+  color: string;
 }
