+++ client/src/interfaces/PaymentType.ts
@@ -1,0 +1,11 @@
+
+export interface PaymentType {
+    id: number;
+    accountId: number;
+    name: string;
+    bgColor: string;
+    fontColor: string;
+    createAt: Date;
+    updateAt: Date;
+    image: string;
+}
